+++ packages/configs/webpack/loaders/css.js
@@ -1,0 +1,14 @@
+const installDependencies = require('../_installer');
+
+const dependencies = ['raw-loader'];
+
+installDependencies(dependencies);
+
+module.exports = {
+  test: /\.css$/,
+  use: [
+    {
+      loader: 'raw-loader',
+    },
+  ],
+};