+++ packages/configs/webpack/CHANGELOG.md
@@ -3,6 +3,58 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-alpha.3](https://github.com/inscriptum/insum/compare/@insum/webpack.config@2.0.0-alpha.2...@insum/webpack.config@2.0.0-alpha.3) (2019-11-07)
+
+**Note:** Version bump only for package @insum/webpack.config
+
+
+
+
+
+# [2.0.0-alpha.2](https://github.com/inscriptum/insum/compare/@insum/webpack.config@2.0.0-alpha.1...@insum/webpack.config@2.0.0-alpha.2) (2019-11-07)
+
+**Note:** Version bump only for package @insum/webpack.config
+
+
+
+
+
+# [2.0.0-alpha.1](https://github.com/inscriptum/insum/compare/@insum/webpack.config@2.0.0-alpha.0...@insum/webpack.config@2.0.0-alpha.1) (2019-11-07)
+
+**Note:** Version bump only for package @insum/webpack.config
+
+
+
+
+
+# [2.0.0-alpha.0](https://github.com/inscriptum/insum/compare/@insum/webpack.config@1.1.0-alpha.0...@insum/webpack.config@2.0.0-alpha.0) (2019-11-07)
+
+
+### Features
+
+* **webpack.config:** 🌟 individual webpack loaders ([b80bbbd](https://github.com/inscriptum/insum/commit/b80bbbd))
+
+
+### BREAKING CHANGES
+
+* **webpack.config:** To use a loader need to import it separately
+
+
+
+
+
+# [1.1.0-alpha.0](https://github.com/inscriptum/insum/compare/@insum/webpack.config@1.0.0...@insum/webpack.config@1.1.0-alpha.0) (2019-11-07)
+
+
+### Features
+
+* **webpack.config:** 🌟 remove loader dependencies from package dependencies ([f2ba20c](https://github.com/inscriptum/insum/commit/f2ba20c))
+* **webpack.config:** 🌟 separate webpack loaders ([b11ebe9](https://github.com/inscriptum/insum/commit/b11ebe9))
+
+
+
+
+
 # [1.0.0](https://github.com/inscriptum/insum/compare/@insum/webpack.config@1.0.0-alpha.0...@insum/webpack.config@1.0.0) (2019-10-08)
 
 **Note:** Version bump only for package @insum/webpack.config